requirements.txt
accelerate==1.10.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
dnspython==2.7.0
faiss-cpu==1.13.0
fastapi==0.115.14
filelock==3.19.1
fsspec==2025.7.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
isort==6.0.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
kernels==0.9.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
openai==1.98.0
packaging==25.0
pillow==12.0.0
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.10.1
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.4
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.2.2
tqdm==4.67.1
transformers==4.56.0
typing_extensions==4.14.1
typing-inspection==0.4.1
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
